Change proctests-test executable name so that it doesn't fail
This test expects that the executable is run in a specific path
and that the executable name starts with "proc_tests".
The first one is already correct but the second one needs to be changed.

This change manages to keep the "-test" postfix, so that the CI keeps
copying the binary.