Be sure to use homebrew libc++ on macOS
It was previously linking to the system libc++.
A new CMake variable has been added, LLVM_INSTALL_PATH, to specify where
the LLVM libraries we want to use are.
By default it searches for libc++ under /usr/local/Cellar/llvm@6/*/lib.

One caveat is that homebrew llvm doesn't provide libc++abi, so the
system one will be used.
It shouldn't be a problem, even though what we would really like is to
compile with AppleClang, but we can't because upstream boost is too old.