Do not require the source directory to create packages
All the needed files should be found in the build directory.

This is mainly for the CI, so that only the build folder is needed
and there's no risk of rebuilding already built binaries because
tar or zip changed timestamps on source files.

The build system used by the CI is also changed to Ninja, because
even when not using the "package" target and calling cpack directly,
it would call the Makefile "preinstall" target.
That requires the source directory to run and we don't want that.
Ninja is in general better and doesn't have any preinstall target,
so cpack can create the package without the source directory.